requirements.txt
asgiref==3.3.1
awscli==1.18.216
botocore==1.19.56
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
Django==3.1.5
docutils==0.15.2
idna==2.10
jmespath==0.10.0
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2

django-cors-middleware==1.5.0
django-extensions==3.1.0
djangorestframework==3.12.2
PyJWT==1.7.1
channels
channels_redis
